Rename: authConstant.js -> auth.js
